apps/backend/chat/gemini_client.py
import os
from google import genai

api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    raise RuntimeError("GEMINI_API_KEY is not set in environment variables.")

client = genai.Client(api_key=api_key)

MODEL_NAME = "gemini-2.5-flash"
