apps/backend/requirements.txt
Django==5.0.6
python-dotenv==1.0.1
google-genai==0.3.0
